Move tagging-related unit tests to new file
